Whitespace-only validation of snipppet fields
The fields "title", "description" and "code" of a the snippet model
are required, yet allow the user to enter whitespace-only content.
This leads to some issues esp. with the title field, which is used
as link text in snippet listings.

This change first adds additional validation to these fields by
checking for content that contains only whitespaces. For snippets
affected by this already in the database the listing now checks
for empty strings and renders them as "Untitled".

Additionally the snippet listing in the admin now includes the
snippet's ID as link-element.